% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awhina-package.R
\docType{package}
\name{awhina-package}
\alias{awhina}
\alias{awhina-package}
\title{awhina: Simplify the Process of Writing R Code at Auckland Council}
\description{
A set of utility functions for Auckland Council R developers to get up and running quickly with the language.
}
\author{
\strong{Maintainer}: Lee Durbin \email{lee.durbin@aucklandcouncil.govt.nz}

Other contributors:
\itemize{
  \item Auckland Council [copyright holder]
}

}
\keyword{internal}
