% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{bind_dataframes}
\alias{bind_dataframes}
\title{Name And Bind Dataframes}
\usage{
bind_dataframes(data, file_names)
}
\arguments{
\item{data}{A list of dataframes.}

\item{file_names}{A character vector of file names from which each dataframe
was derived, where the position of each file name in the character vector
matches the position of the dataframe in the list.}
}
\value{
A tibble.
}
\description{
Name And Bind Dataframes
}
