% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{check_environment_variable}
\alias{check_environment_variable}
\title{Check if the user has their environment variables set up}
\usage{
check_environment_variable(variable_names)
}
\arguments{
\item{variable_names}{a character vector of environment variable names in the .Renviron file}
}
\value{
a stylised message in the console indicating failure or success
}
\description{
Given a list of environment variable names, returns a warning in the console if
the user does not have these set in their .Renviron file
}
