% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{check_file_properties}
\alias{check_file_properties}
\title{Check File Properties}
\usage{
check_file_properties(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A list containing two logical items
}
\description{
Check file properties and print a warning if the check fails.
}
