% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{check_renviron}
\alias{check_renviron}
\title{Check that a .Renviron exists}
\usage{
check_renviron(renviron_path)
}
\arguments{
\item{renviron_path}{Path to the .Renviron file.}
}
\value{
Scalar logical indicating if the file was read successfully. Returned invisibly.
}
\description{
Check that a .Renviron exists
}
