% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{connect_to_sharepoint}
\alias{connect_to_sharepoint}
\title{Connect To A SharePoint Site}
\usage{
connect_to_sharepoint(site_name)
}
\arguments{
\item{site_name}{The name of the SharePoint site to which you wish to
connect, which can be found in the site's homepage URL.}
}
\value{
An R6 object of class ms_site.
}
\description{
Connect To A SharePoint Site
}
