% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{create_environment_variable}
\alias{create_environment_variable}
\title{Create a .Renviron file to help point to your SharePoint document library}
\usage{
create_environment_variable(value)
}
\arguments{
\item{value}{A character string containing the path to your SharePoint Document Library
as sync'd to your local machine}
}
\value{
A stylised message in the console
}
\description{
Given a path to your locally-sync'd SharePoint Document Library, create a .Renviron
file in the project root to store that information for future reference.
}
\examples{
create_environment_variable("C:/path/to/sharepoint/directory")
}
