% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fy_to_date}
\alias{fy_to_date}
\title{Convert a Financial Year To A Date}
\usage{
fy_to_date(fy, month, starts, day = "01")
}
\arguments{
\item{fy}{Character string of Financial Year, expressed either as "FYYY" or
simply "YY".}

\item{month}{Name of month expressed as unabbreviated character string (e.g.
"August").}

\item{starts}{Name of the month in which the Financial Year starts, expressed
as unabbreviated character string (e.g. "July").}

\item{day}{Number expressing the day of the month, expressed as a character
vector. Default is "01".}
}
\value{
An S3 object of class \if{html}{\out{<Date>}}.
}
\description{
Given a Financial Year (e.g. "FY24) and a month, convert it to
an actual date.
}
\examples{
fy_to_date("FY24", "August", "July")
}
