% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_data_alias_table}
\alias{get_data_alias_table}
\title{Get Alias Table Data From MS List}
\usage{
get_data_alias_table()
}
\value{
A dataframe containing five columns from the alias table
}
\description{
Read data from the Libraries Subscription Databases Alias Table
MS List and return it as a dataframe.
}
