% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_data_register}
\alias{get_data_register}
\title{Get Register Data From MS List}
\usage{
get_data_register()
}
\value{
A dataframe containing data from the Libraries Subscription Databases
Register
}
\description{
Read data from the Libraries Subscription Databases Register MS
List and return it as a dataframe.
}
