% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{get_excel_file}
\alias{get_excel_file}
\title{Efficiently read an Excel file}
\usage{
get_excel_file(
  filename,
  sheetname = 1,
  skip_rows = 0,
  path = get_file_storage_path("MY_SHAREPOINT_FILES"),
  file_extension = ".xlsx"
)
}
\arguments{
\item{filename}{The name of the file, excluding the file extension (e.g. .xlsx)}

\item{sheetname}{The name or position of the sheet you'd like to read. Defaults to position 1.}

\item{skip_rows}{The number of rows to skip before any data is read from the sheet. Default to zero rows skipped.}

\item{path}{The full directory path to the file. Default to the file path you stored in .Renviron as MY_SHAREPOINT_FILES.}

\item{file_extension}{The file extension of the file. Default to .xlsx files.}
}
\value{
A tibble
}
\description{
Pass the name of an Excel file and instantly read it. Optionally specify a sheet name and other parameters.
}
