% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_file_name}
\alias{get_file_name}
\title{Get The Names Of Files In A Folder}
\usage{
get_file_name(file_path = tere::get_file_storage_path(), folder_name)
}
\arguments{
\item{file_path}{The path of the file.}

\item{folder_name}{The name of the folder housing the files.}
}
\value{
File names as a named fs_path character vector.
}
\description{
Given a file path and a folder name, return the name of any
files in that folder.
}
