% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{get_file_properties}
\alias{get_file_properties}
\title{Get File Properties}
\usage{
get_file_properties(folder_path)
}
\arguments{
\item{folder_path}{The path of the folder}
}
\value{
A dataframe containing the files and their properties
}
\description{
Get the properties of files within a folder
}
