% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{get_list_items}
\alias{get_list_items}
\title{Retrieve SharePoint List Items}
\usage{
get_list_items(site_name, list_name)
}
\arguments{
\item{site_name}{The name of the SharePoint site to which you wish to
connect, which can be found in the site's homepage URL.}

\item{list_name}{The name of the SharePoint List you wish to retrieve.}
}
\value{
A data frame.
}
\description{
Retrieve SharePoint List Items
}
