% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{get_packages}
\alias{get_packages}
\title{Install any packages that are not already installed}
\usage{
get_packages(packages)
}
\arguments{
\item{packages}{a character vector of package names.}
}
\value{
A stylised message in the console.
}
\description{
Compare a list of package names (including utils) against installed packages, then install any as needed.
}
\examples{
get_packages(c("dplyr", "lubridate"))
}
