% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{get_renviron}
\alias{get_renviron}
\title{Obtain the values of environment variables}
\usage{
get_renviron(var_name)
}
\arguments{
\item{var_name}{Character vector of environment variable name(s).}
}
\value{
A vector of the same length as var_name.
}
\description{
Obtain the values of environment variables
}
