% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_file}
\alias{read_file}
\title{Read Files}
\usage{
read_file(
  file_names,
  file_path = paste0(tere::get_file_storage_path(), folder_name),
  file_type = c("excel", "csv", "txt"),
  file_extension = c(".xlsx", ".xls", ".csv"),
  sheet = 1,
  skip = 0
)
}
\arguments{
\item{file_path}{The path of the file(s) to be read}

\item{file_type}{The format of the file(s) to be read.}

\item{sheet}{The sheet number in the excel file(s) to be read. Defaults to
the first sheet.}

\item{skip}{The number of rows to skip before reading in the data.}

\item{file_name}{The name of the file(s) to be read.}
}
\value{
Data from the file(s)
}
\description{
Read Files
}
