% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tere-package.R
\docType{package}
\name{tere-package}
\alias{tere}
\alias{tere-package}
\title{tere: Simplify the Process of Writing R Code at Auckland Council}
\description{
A set of utility functions for Auckland Council R developers to get up and running quickly with the language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lddurbinac.github.io/tere/}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{lee.durbin@aucklandcouncil.govt.nz}

Other contributors:
\itemize{
  \item Auckland Council [copyright holder]
}

}
\keyword{internal}
