% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{update_list_items}
\alias{update_list_items}
\title{Update SharePoint List Items}
\usage{
update_list_items(site_name, list_name, item_ids, ...)
}
\arguments{
\item{site_name}{Name of SharePoint site to which you want to connect.}

\item{list_name}{Name of the List to which you are connecting.}

\item{item_ids}{Character vector containing a list of IDs to SharePoint List items.}

\item{...}{Named character vector.}
}
\value{
Message to the console
}
\description{
Update SharePoint List Items
}
