% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{check_environment_variable}
\alias{check_environment_variable}
\title{Check if the user has their environment variables set up}
\usage{
check_environment_variable(variable_names)
}
\arguments{
\item{variable_names}{a character vector of environment variable names in the .Renviron file.}
}
\value{
FALSE if any of the environment variables are not found, otherwise TRUE.
}
\description{
Given a list of environment variable names, return FALSE if
the user does not have any of  these set in their .Renviron file.
}
