% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{create_environment_variable}
\alias{create_environment_variable}
\title{Create a .Renviron file that points to your locally-synchronised SharePoint Document Library}
\usage{
create_environment_variable(dir_ref)
}
\arguments{
\item{dir_ref}{A name for your SharePoint Document Library, using only capital letters and no spaces}
}
\value{
A stylised message in the console
}
\description{
Given a path to your locally-sync'd SharePoint Document Library, create a .Renviron
file in the project root to store that information for future reference.
}
\examples{
\dontrun {create_environment_variable("MY_SHAREPOINT_FILES")}
}
