% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started.R
\name{get_file_storage_path}
\alias{get_file_storage_path}
\title{Retrieve the full path to the local SharePoint directory where your team's files are stored}
\usage{
get_file_storage_path(
  dir_ref = "MY_SHAREPOINT_FILES",
  renviron_path = here::here(".Renviron")
)
}
\arguments{
\item{dir_ref}{The name you assigned for your SharePoint Document Library when you called create_environment_variable(), defaults to "MY_SHAREPOINT_FILES".}

\item{renviron_path}{The path to the .Renviron file, defaults to your project's home directory.}
}
\value{
A string containing a Windows directory path, or else an error message.
}
\description{
Retrieve the full path to the local SharePoint directory where your team's files are stored
}
\examples{
\dontrun{get_file_storage_path("MY_SHAREPOINT_FILES")}
}
