% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{update_list_items}
\alias{update_list_items}
\title{Update SharePoint List Items}
\usage{
update_list_items(sharepoint_list, item_ids, ...)
}
\arguments{
\item{sharepoint_list}{Connection to a SharePoint List}

\item{item_ids}{Character vector containing a list of IDs to SharePoint List items}

\item{...}{Named character vector}
}
\value{
Message to the console
}
\description{
Update SharePoint List Items
}
